--- pages/EntrancePage copy.js
@@ -1,10 +1,0 @@
-import React from 'react'
-import { StyleSheet, View, Text, Button } from 'react-native';
-
-const Entrance = () => {
-    return (
-        <Text>Hello this is test</Text>
-    )
-}
-
-export default Entrance
