+++ pages/PopularPage.js
@@ -1,0 +1,15 @@
+import React, { useState } from 'react';
+import { StyleSheet, Text, View, Button } from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
+
+const MoviesScreen = () => {
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Movies Screen</Text>
+    </View>
+  )
+}
+
+export default MoviesScreen
+
