+++ pages/TopRatedPage.js
@@ -1,0 +1,17 @@
+import React, { useState } from 'react';
+import { StyleSheet, Text, View, Button } from 'react-native';
+import { NavigationContainer } from '@react-navigation/native';
+import { createStackNavigator } from '@react-navigation/stack';
+
+const MusicScreen = () => {
+
+
+  return (
+    <View style={{ flex: 1, alignItems: 'center', justifyContent: 'center' }}>
+      <Text>Music Screen</Text>
+    </View>
+  )
+}
+
+export default MusicScreen
+
